requirements.txt
python-telegram-bot==1.8
MySQL-python==1.2.5
pygeocoder==1.2.5